MVP/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.1
alembic==1.13.1
python-dotenv==1.0.0
redis==5.0.1
pydantic-settings==2.1.0
python-rapidjson==1.6
celery==5.3.6
redis-celery==5.3.6
prometheus-fastapi-instrumentator==5.12.0
sentry-sdk==1.38.0
pandas==2.1.4
openpyxl==3.1.2
python-magic==0.4.27
aiohttp==3.9.1
aioredis==2.0.1
fastapi-sse==0.1.0
reportlab==4.0.8
python-magic-bin==0.4.14
sentry-sdk[flask]==1.38.0
slack-sdk==3.21.3
pagerduty-sdk==1.10.0
websockets==12.0
python-socketio==5.10.0
jinja2==3.1.2
weasyprint==60.1
python-pptx==0.6.21
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.17.0
pypng==0.20230715.0
python-dotenv==1.0.0
boto3==1.29.1
schedule==1.2.1
uvloop==0.18.0